+++ src/lib/components/Preview.svelte
@@ -8,8 +8,14 @@
 	export let preview = '';
 
 	const renderWhiteSpace = (text: string) => {
+		const maxLength = 150;
+		if (text.length > maxLength) {
+			throw new RangeError(
+				`Input exceeds maximum length of ${maxLength} characters. A good practice is to keep this text under 90 characters.`
+			);
+		}
 		const whiteSpaceCodes = '\xa0\u200C\u200B\u200D\u200E\u200F\uFEFF';
-		return whiteSpaceCodes.repeat(150 - text.length);
+		return whiteSpaceCodes.repeat(maxLength - text.length);
 	};
 </script>